Debounce note content changes
Currently all data is stored in a single JSON blob in LocalStorage. This
is not very efficient, but otherwise fine for now. But, serializing all
this data on every keypress is problematic, as it makes typing feel very
sluggish.

Debouncing the change events prevents this and is fine in general; we
don't need to persist the change on every keypress.