Allow ids to pre provided on the API during create
Technically that's a bit weird because these are POST calls, but it
provides an opportunity to have clients generate ids for offline
resource creation. A PUT would be more appropriate in this case, but
with the short nanoids that are used a PUT API could lead to data loss
in case of a collision.