Reduce expensive draft.js content diffing
By having React re-create the whole editor when a content block changes.
We can know cheaply if a content block has changed by looking at it's
updatedAt value. This way, we don't need componentWillReceiveProps etc.

This does mean that the editor is completely recreated if a note is
updated from the server, but that's ok as it's difficult to retain
Draft.js cursor position etc when updating from scratch.